+++ searchbase/lib/searchbase.dart
@@ -4,3 +4,4 @@ export 'src/constants.dart';
 export 'src/results.dart';
 export 'src/types.dart';
 export 'src/aggregations.dart';
+export 'src/searchstatecontroller.dart';